Upgrade to net6 and cleanup (#26)

+++ src/Web/Controllers/v1/CustomersController.cs
@@ -6,6 +6,7 @@ using SomeBasicEFApp.Web.Data;
 using SomeBasicEFApp.Web.Entities;
 using SomeBasicEFApp.Web.Models;
 using SomeBasicEFApp.Web.ValueTypes;
+using Microsoft.AspNetCore.Http;
 
 namespace SomeBasicEFApp.Web.Controllers.v1
 {
@@ -31,7 +32,10 @@ namespace SomeBasicEFApp.Web.Controllers.v1
 
         // GET: Customers/Details/5
         [HttpGet("{id}")]
-        [Produces(typeof(CustomerModel))]
+        [ProducesResponseType(typeof(CustomerModel),StatusCodes.Status200OK)]
+        [ProducesResponseType(typeof(void),StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(typeof(void),StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType()]
         public async Task<IActionResult> Details(CustomerId? id)
         {
             if (id == null)
@@ -61,6 +65,9 @@ namespace SomeBasicEFApp.Web.Controllers.v1
 
         [HttpPut("{id}")]
         [Produces(typeof(CustomerModel))]
+        [ProducesResponseType(typeof(CustomerModel),StatusCodes.Status200OK)]
+        [ProducesResponseType(typeof(void),StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType]
         public async Task<IActionResult> Edit(CustomerId id, [FromBody] EditCustomer model)
         {
             var customer = await _context.GetCustomerAsync(id);
@@ -75,9 +82,13 @@ namespace SomeBasicEFApp.Web.Controllers.v1
         // POST: Customers/Delete/5
         [HttpDelete("{id}")]
         [Produces(typeof(void))]
+        [ProducesResponseType(typeof(void),StatusCodes.Status200OK)]
+        [ProducesResponseType(typeof(void),StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType]
         public async Task<IActionResult> DeleteConfirmed(CustomerId id)
         {
             var customer = await _context.Customers.SingleOrDefaultAsync(m => m.Id == id);
+            if (customer is null) return NotFound();
             _context.Customers.Remove(customer);
             await _context.SaveChangesAsync();
             return Ok();
